Move info commands to slashies
